UI and logic improved
