+++ backend/contacts.json
@@ -1,7 +1,12 @@
 [
   {
-    "nome": "Babbo",
-    "cognome": " ",
-    "numero": "393313376465"
+    "nome": "Michela",
+    "cognome": "Bartolini",
+    "numero": "393497181104"
+  },
+  {
+    "nome": "Eszter",
+    "cognome": "Erasmus",
+    "numero": "36209825327"
   }
 ]
